+++ ruff-lint-strict.toml
@@ -9,7 +9,7 @@ exclude = [
     "alphadia/__*__.py",
     "alphadia/cli.py",
     "alphadia/exceptions.py",
-    "alphadia/fdr.py",
+    "alphadia/fdr/**.py",
     "alphadia/plexscoring/features/**.py",
     "alphadia/fragcomp.py",
     "alphadia/grouping.py",
@@ -30,7 +30,6 @@ exclude = [
     "alphadia/_fdrx/stats.py",
     "alphadia/_fdrx/plotting.py",
     "alphadia/_fdrx/base.py",
-    "alphadia/fdr_/**.py",
     "alphadia/numba/*.py",
     "alphadia/peakgroup/*.py",
     "alphadia/plotting/*.py",
